requirements.txt
customtkinter>=5.2.2
psutil>=5.9.8
GPUtil
pystray
Pillow
customtkinter
psutil
